Moved polls state to PollList and changed filter to only Open polls
